Add the build script
- Coming from iitc-ce
  https://github.com/IITC-CE/ingress-intel-total-conversion
- Adapted by LeJeu
  https://github.com/le-jeu/iitc-plugins
- Modified so building is also working in windows shell.

The build script is build_plugin.py

buildsettings.py is used to define build profile.
A dev profile dw235dev is defined in it.
A release profile dw235 is commented out,
the url needs to be set where release artifacts are available.

Running it :
python3 build_plugin.py --out-dir dist dw235dev "plugins/*.js"

dist is the path to the out directory
dw235dev is the name of the profile set in buildsettings.py
"plugins/*.js" defines what plugin files to build